Break out ITest cleanup into another base file, move some slower tests to only do one setup + cleanup
