+++ src/test/java/com/basho/riak/client/api/commands/itest/ITestBucketKeyMapReduce.java
@@ -21,6 +21,7 @@ import com.basho.riak.client.api.commands.buckets.StoreBucketProperties;
 import com.basho.riak.client.api.commands.kv.StoreValue;
 import com.basho.riak.client.api.commands.mapreduce.BucketKeyMapReduce;
 import com.basho.riak.client.api.commands.mapreduce.MapReduce;
+import com.basho.riak.client.core.operations.itest.ITestBase;
 import com.basho.riak.client.core.query.Location;
 import com.basho.riak.client.core.query.Namespace;
 import com.basho.riak.client.core.query.RiakObject;
@@ -29,52 +30,59 @@ import com.basho.riak.client.core.util.BinaryValue;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
-import org.junit.Test;
+import org.junit.*;
 
 import java.io.IOException;
 import java.util.Map;
 import java.util.concurrent.ExecutionException;
-import org.junit.After;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
-import org.junit.Assume;
-import org.junit.Before;
 
 /**
  * @author Brian Roach <roach at basho dot com>
  * @since 2.0
  */
-public class ITestBucketKeyMapReduce extends ITestAutoCleanupBase
+public class ITestBucketKeyMapReduce extends ITestBase
 {
+    private final static RiakClient client = new RiakClient(cluster);
+    private final static String mrBucketName = "ITestBucketKeyMapReduce";
 
-    RiakClient client = new RiakClient(cluster);
-    static final String mrBucket = "mr_bucket";
-    
-    @Before
-    public void changeBucketProps() throws ExecutionException, InterruptedException
+    @BeforeClass
+    public static void setup() throws InterruptedException, ExecutionException
     {
+        changeBucketProps();
+        initValues(Namespace.DEFAULT_BUCKET_TYPE.toString());
         if (testBucketType)
         {
-            Namespace ns = new Namespace(bucketType.toString(), mrBucket);
-            StoreBucketProperties op = new StoreBucketProperties.Builder(ns).withAllowMulti(false).build();
-            client.execute(op);
+            initValues(mapReduceBucketType.toString());
         }
     }
-    
-    @After
-    public void clearMrBucket() throws InterruptedException, ExecutionException
+
+    @AfterClass
+    public static void tearDown() throws ExecutionException, InterruptedException
     {
-        Namespace ns = new Namespace(Namespace.DEFAULT_BUCKET_TYPE, mrBucket);
+        Namespace ns = new Namespace(mrBucketName);
         resetAndEmptyBucket(ns);
+
         if (testBucketType)
         {
-            ns = new Namespace(bucketType.toString(), mrBucket);
+            ns = new Namespace(mapReduceBucketType.toString(), mrBucketName);
             resetAndEmptyBucket(ns);
         }
     }
 
-    private void initValues(String bucketType) throws ExecutionException, InterruptedException
+    private static void changeBucketProps() throws ExecutionException, InterruptedException
+    {
+        if (testBucketType)
+        {
+            Namespace ns = new Namespace(mapReduceBucketType.toString(), mrBucketName);
+            StoreBucketProperties op = new StoreBucketProperties.Builder(ns).withAllowMulti(false).build();
+            client.execute(op);
+        }
+    }
+
+    private static void initValues(String bucketType) throws ExecutionException, InterruptedException
     {
         RiakObject obj = new RiakObject();
 
@@ -83,7 +91,7 @@ public class ITestBucketKeyMapReduce extends ITestAutoCleanupBase
                 "book her sister was reading, but it had no pictures or conversations in " +
                 "it, 'and what is the use of a book,' thought Alice 'without pictures or " +
                 "conversation?'"));
-        Namespace ns = new Namespace(bucketType, mrBucket);
+        Namespace ns = new Namespace(bucketType, mrBucketName);
         Location location = new Location(ns, "p1");
         client.execute(new StoreValue.Builder(obj).withLocation(location).build());
 
@@ -116,14 +124,12 @@ public class ITestBucketKeyMapReduce extends ITestAutoCleanupBase
     public void JsBucketKeyMRTestType() throws InterruptedException, ExecutionException, IOException
     {
         Assume.assumeTrue(testBucketType);
-        JsBucketKeyMR(bucketType.toString());
+        JsBucketKeyMR(mapReduceBucketType.toString());
     }
     
     private void JsBucketKeyMR(String bucketType) throws InterruptedException, ExecutionException, IOException
     {
-        initValues(bucketType);
-        
-        Namespace ns = new Namespace(bucketType, mrBucket);
+        Namespace ns = new Namespace(bucketType, mrBucketName);
         MapReduce mr = new BucketKeyMapReduce.Builder()
                 .withLocation(new Location(ns, "p1"))
                 .withLocation(new Location(ns, "p2"))
@@ -188,13 +194,12 @@ public class ITestBucketKeyMapReduce extends ITestAutoCleanupBase
     public void erlangBucketKeyMRTestType() throws ExecutionException, InterruptedException
     {
         Assume.assumeTrue(testBucketType);
-        erlangBucketKeyMR(bucketType.toString());
+        erlangBucketKeyMR(mapReduceBucketType.toString());
     }
     
     private void erlangBucketKeyMR(String bucketType) throws ExecutionException, InterruptedException
     {
-        initValues(bucketType);
-        Namespace ns = new Namespace(bucketType, mrBucket);
+        Namespace ns = new Namespace(bucketType, mrBucketName);
         
         MapReduce mr = new BucketKeyMapReduce.Builder()
                 .withLocation(new Location(ns, "p1"))
