+++ src/test/java/com/basho/riak/client/core/operations/itest/ITestAutoCleanupBase.java
@@ -47,80 +47,6 @@ public class ITestAutoCleanupBase extends ITestBase
             resetAndEmptyBucket(defaultNamespace());
         }
     }
-    
-    public static void resetAndEmptyBucket(BinaryValue name) throws InterruptedException, ExecutionException
-    {
-        resetAndEmptyBucket(new Namespace(Namespace.DEFAULT_BUCKET_TYPE, name.toString()));
-    }
-
-    public static void resetAndEmptyBucket(Namespace namespace) throws InterruptedException, ExecutionException
-    {
-        ListKeysOperation.Builder keysOpBuilder = new ListKeysOperation.Builder(namespace);
-        
-        ListKeysOperation keysOp = keysOpBuilder.build();
-        cluster.execute(keysOp);
-        List<BinaryValue> keyList = keysOp.get().getKeys();
-        final Semaphore semaphore = new Semaphore(NUMBER_OF_PARALLEL_REQUESTS);
-        final CountDownLatch latch = new CountDownLatch(keyList.size());
-        
-        RiakFutureListener<Void, Location> listener = new RiakFutureListener<Void, Location>() {
-            @Override
-            public void handle(RiakFuture<Void, Location> f)
-            {
-                try
-                {
-                    f.get();
-                }
-                catch (Exception ex)
-                {
-                    if (ex instanceof RuntimeException)
-                    {
-                        throw (RuntimeException)ex;
-                    }
-                    throw new RuntimeException(ex);
-                }
-                semaphore.release();
-                latch.countDown();
-            }
-            
-        };
-        
-        for (BinaryValue k : keyList)
-        {
-            Location location = new Location(namespace, k);
-            DeleteOperation delOp = new DeleteOperation.Builder(location).withRw(3).build();
-            delOp.addListener(listener);
-            semaphore.acquire();
-            cluster.execute(delOp);
-        }
-
-        latch.await();
-
-        ResetBucketPropsOperation.Builder resetOpBuilder = 
-            new ResetBucketPropsOperation.Builder(namespace);
-        
-        ResetBucketPropsOperation resetOp = resetOpBuilder.build();
-        cluster.execute(resetOp);
-        resetOp.get();
-
-    }
-    
-    public static boolean assureIndexExists(String indexName) throws InterruptedException
-    {
-        for (int x = 0; x < 5; x++)
-        {
-            Thread.sleep(2000);
-            YzFetchIndexOperation fetch = new YzFetchIndexOperation.Builder().withIndexName(indexName).build();
-            cluster.execute(fetch);
-            fetch.await();
-            if (fetch.isSuccess())
-            {
-                return true;
-            }
-        }
-        
-        return false;
-    }
 
     public static Namespace defaultNamespace()
     {
